Merge in RP/react-ui-components from feature/some-feature to master

* commit 'ad79c0b233d8d1b80dc54f0214d5d0298f708614':
  version updated
  v0.7.4
  the logic select erase item edited
  versions updated
  v0.7.3
  the Input event handler edited
  the inputRef added into the Input
  versions updated
  v0.7.2
  some validated fixed
  versions updated
  v0.7.1
  the select search some improved
  snapshots updated
  versions updated
  v0.7.0
  The error styles improved input textfield datepicker
  The clean flag into the datepicker added

+++ packages/datepicker/src/index.tsx
@@ -42,9 +42,20 @@ const Datepicker: React.FunctionComponent<IDatepicker> = (props: IDatepicker) =>
   const [numberDayInWeek, setNumberDayInWeek] = useState(dateParsed?.getNumberDayInWeek());
   const [countDaysIsMonth, setCountDaysIsMonth] = useState(dateParsed?.getCountDaysInMonth());
   const [isError, setIsError] = useState(false);
+  const [isMinDateError, setIsMinDateError] = useState(false);
+  const [isMaxDateError, setIsMaxDateError] = useState(false);
+
+  const inputRef: any = useRef() as React.MutableRefObject<HTMLInputElement>;
+
+  const minDate = new DateParser(props.minDate);
+  const maxDate = new DateParser(props.maxDate);
+  const minDateMilliseconds: number = minDate?.getTimestamp();
+  const maxDateMilliseconds: number = maxDate?.getTimestamp();
 
   const onInputDelete = (name: string) => {
-    props?.onRemove(name, null);
+    if (props?.onRemove) {
+      props?.onRemove(name, null);
+    }
     setIsExistValue(false);
     setIsVisibleList(false);
     setValue('');
@@ -56,6 +67,14 @@ const Datepicker: React.FunctionComponent<IDatepicker> = (props: IDatepicker) =>
       const listElement: any = dateRef?.current;
       if (listElement) {
         onDatesContainerClose(searchDomChildElement(listElement, element), evt);
+      }
+    }
+    if (inputRef) {
+      const input = inputRef?.current;
+      if (element !== input) {
+        if (input?.blur) {
+          input.blur();
+        }
       }
     }
   };
@@ -120,13 +139,28 @@ const Datepicker: React.FunctionComponent<IDatepicker> = (props: IDatepicker) =>
     const regExpr: RegExp = new RegExp('(?<day>[0-9]{2})(?<month>[0-9]{2})(?<year>[0-9]{4})');
     const datePartitioned: { day: string, month: string, year: string } =
       element?.value !== null ? element?.value?.replaceAll('.', '')?.match(regExpr)?.groups : null;
-    const valueParsed: string = `${datePartitioned.day}.${datePartitioned.month}.${datePartitioned.year}`;
+    const valueParsed: string = datePartitioned
+      ? `${datePartitioned?.day}.${datePartitioned?.month}.${datePartitioned?.year}`
+      : null;
     setValue(valueParsed);
     dateParsed = new DateParser(valueParsed);
-    if (props.onChange && dateParsed.checkIsValidateDate()) {
+    const dateParsedMilliseconds: number = dateParsed.getTimestamp();
+    if (
+      props.onChange &&
+      dateParsed.checkIsValidateDate() &&
+      ((minDateMilliseconds && dateParsedMilliseconds >= minDateMilliseconds) || true) &&
+      ((maxDateMilliseconds && dateParsedMilliseconds <= maxDateMilliseconds) || true)
+    ) {
       props.onChange(props.name, valueParsed, true);
       setIsError(false);
+      setIsMinDateError(false);
+      setIsMaxDateError(false);
     } else {
+      if (dateParsedMilliseconds < minDateMilliseconds) {
+        setIsMinDateError(true);
+      } else if (dateParsedMilliseconds > maxDateMilliseconds) {
+        setIsMaxDateError(true);
+      }
       setIsError(true);
     }
   };
@@ -147,7 +181,7 @@ const Datepicker: React.FunctionComponent<IDatepicker> = (props: IDatepicker) =>
 
   const months: Array<string> = props.locale === Locales.Ru || !props?.locale ? monthsElementRu : monthsElementEn;
   const years: Array<string> = new Array(100)
-    .fill(currentYearNumber ?? 0 - 50)
+    .fill((currentYearNumber ?? 0) - 50)
     .map((element: number, index: number) => (element + index)?.toString());
 
   const componentThemed: any = (theme: Itheme) => {
@@ -216,7 +250,19 @@ const Datepicker: React.FunctionComponent<IDatepicker> = (props: IDatepicker) =>
     const yearTitle: string = props?.locale === Locales.Ru || !props?.locale ? 'Год' : 'Year';
 
     const textMessageError: string = props?.locale === Locales.Ru ? 'Дата не валидна' : 'Date is not valid';
-    const textMessage = isError ? textMessageError : props?.textMessage;
+
+    const minDateMessage: string =
+      props?.locale === Locales.Ru ? 'Дата меньше допустимой' : 'Date is less than allowed';
+    const maxDateMessage: string =
+      props?.locale === Locales.Ru ? 'Дата больше допустимой' : 'Date is more then allowed';
+
+    const textMessage = isError
+      ? isMinDateError
+        ? minDateMessage
+        : isMaxDateError
+        ? maxDateMessage
+        : textMessageError
+      : props?.textMessage;
 
     const currentYearNumberString: string = currentYearNumber.toString();
 
@@ -237,6 +283,7 @@ const Datepicker: React.FunctionComponent<IDatepicker> = (props: IDatepicker) =>
                 fontWeight={props?.fontWeight}
                 isDisabled={props.disabled}
                 fontFamily={props?.fontFamily || theme?.fontFamily}
+                error={isError}
               >
                 {props?.label}
               </Label>
@@ -266,6 +313,7 @@ const Datepicker: React.FunctionComponent<IDatepicker> = (props: IDatepicker) =>
             isReadOnly={props?.isReadOnly}
             isNotUseDebounce={true}
             error={isError}
+            inputRef={inputRef}
           />
         </DatepickerHeader>
         {isVisibleList && (
